Implement interim account delete flow

+++ app/Http/Controllers/UserPageController.php
@@ -14,21 +14,22 @@ class UserPageController extends Controller
     public function show(string $username) {
         $user = User::where("name", $username)->first();
 		$user->profile_html = $this->sanitise_html($user->profile_html);
-		$user->avatar_url = Storage::url($user->avatar);
+		$user->avatar_url = $user->avatar ? Storage::url($user->avatar) : "/images/user.png";
         return view("profile.show", ["user" => $user]);
     }
 
 	public function edit() {
 		$user = Auth::user();
-		$user->avatar_url = Storage::url($user->avatar);
+		$user->avatar_url = $user->avatar ? Storage::url($user->avatar) : "/images/user.png";
 		return view("profile.html.edit", ["user" => $user]);
 	}
 
 	public function update(UserPageUpdateRequest $request, UploadService $uploadService) {
+		// Validate HTML
 		$profile_html = $request->validated()["profile_html"];
 
-		$uploadService->delete($request->user()->avatar);
-		$avatar = $uploadService->upload($request->avatar, "avatars/".$request->user()->name);
+		if($request->user()->avatar) $uploadService->delete($request->user()->avatar);
+		$avatar = $uploadService->upload($request->validated()["avatar"], "avatars/".$request->user()->name);
 		$uploadService->resizeToFit($avatar, 300);
 
 		$request->user()->profile_html = $profile_html;
